refactor: move vtable logic into own file
