+++ packages/lix-docs/rspress.config.ts
@@ -125,6 +125,7 @@ export default defineConfig({
             },
             { text: "Versions", link: "/guide/concepts/versions" },
             { text: "Discussions", link: "/guide/concepts/discussions" },
+            { text: "Key-Value", link: "/guide/concepts/key-value" },
           ],
         },
         {
